requirements.txt
streamlit>=1.32.0
google-auth>=2.49.0
google-auth-httplib2>=0.3.0
google-api-python-client>=2.194.0
google-genai>=1.0.0
Pillow>=10.0.0
